app/requirements.txt
fastapi
pydantic
uvicorn
paho-mqtt
requests
